bussy update

+++ orderHashmap.cpp
@@ -1,8 +1,8 @@
 #include "orderHashmap.h"
 #include <iostream>
 
-int orderHashmap::hash(int key) {
-    return (key % (int) this->capacity);
+long orderHashmap::hash(int key) {
+    return (key % (long) this->capacity);
 }
 
 void orderHashmap::rehash() {
